requirements.txt
# Core dependencies
python-dateutil>=2.8.2
PyYAML>=6.0
rtyaml>=1.0.0
lxml>=4.9.3
requests>=2.28.2
urllib3>=1.26.15
certifi>=2022.12.7

# Data collection and processing
congress>=3.1.0
scrapelib>=2.1.0
rdflib>=6.0.0
SPARQLWrapper>=2.0.0

# Database
sqlite3>=2.6.0

# AI / ML
google-generativeai>=0.3.0

# Date/time handling
pytz>=2023.3
tzlocal>=4.3.1

# CLI enhancements
colorama>=0.4.6
tqdm>=4.66.1

# Development and debugging
ipython>=8.12.0
pyflakes>=3.0.1

# Social media integration
tweepy>=4.14.0
requests-oauthlib>=1.3.1
oauthlib>=3.2.2

# Web scraping and parsing
cssselect>=1.2.0
beautifulsoup4>=4.12.2

# Logging enhancements
rich>=13.4.2

# Note: This project requires Cursor IDE (https://cursor.sh/) for development
# Some dependencies may need to be adjusted based on your operating system 